Increased icon size, moved SettingsFragment to it's own class
